move components folder and tests files out of pages to prevent build errors
