+++ pages/_app.tsx
@@ -1,6 +1,6 @@
 import '@/styles/globals.css';
 import type { AppProps } from 'next/app';
-import NavBar from './components/navbar';
+import NavBar from '../components/navbar';
 
 export default function App({ Component, pageProps }: AppProps) {
   return (
